Fix Memory ReAllocation Bug
The program has two arrays Wall_Polygon and Cut_Polygon. These both store Polygon_Node structs that used to store pointer references to nodes in the other array.

The error came from when one of the two TArrays reallocated their place in memory, causing the pointers to be referencing undefined values.

Fixed problem by storing array indices corresponding to the other TArray nodes.

Problems could still occur if nodes are inserted not at the end of the TArray, offsetting the index count. However in current implementation this is not a problem.