Separate Vertex from CLL container
Now vertices are initialized as a struct then added into the Circular Linked List as a Class that is created in heap memory. This allows the user of Polygon.h to not have responsibility of deleting memory related to vertices. However those vertex data can still be accessed either through the return value of the Add or by a iterator.