Fix Vertex Skipping in Wall_Cutter::Add_Intercepts
Found intercepts were skipped in the iteration, this was done to prevent the same intercepts from continuously being added. This caused vertices to be skipped if two intercepts were on a line between two non intercept points. The second intercept would not be found and the iterator would jump to the second of the two non intercept points.

This fix causes a known inefficiency where every intercept added will be found as a intercept but then caught by the check for repeat line. A change to when points are added or a deeper understanding of moving pointer references would help fix this issue!