Fix Polygon Triangulation Error with PointInside Check
When checking if a point was inside a polygon. If the point was perfectly aligned on the infinite line an edge makes, then that point was projecting on that infinite line and seemed at a distance of 0 from the line. Now we check if the Dot product is longer than the line edges length.